requirements.txt
hyperliquid-python-sdk==0.18.0
python-dotenv==1.1.1
eth-account==0.13.7
requests==2.32.5